+++ frame/support/src/traits/tokens/fungible/conformance_tests/mod.rs
@@ -1,1 +1,1 @@
-pub mod inspect_mutate;
+pub mod regular;
